feat(profile): 프로필 이미지를 BunnyCDN 저장소에 저장
업로드된 이미지를 본 서버 디스크 대신 BunnyCDN 스토리지 존에 올린다.
multer를 memoryStorage로 바꿔 파일이 디스크에 닿지 않게 하고, sharp가
버퍼를 그대로 처리한 뒤 결과만 저장소로 PUT 한다. 저장되는 주소는
기존 Pull Zone(config.project.cdn) 아래 profiles/ 경로다.

교체 시 이전 파일 삭제는 우리가 올린 파일에만 적용한다. 주소 접두사와
파일명 규칙으로 판별해 Bunny면 Storage DELETE, 전환 전 업로드분이면
디스크에서 지운다. 기본 이미지나 외부 아바타는 대상에서 빠진다.

업로드 후 백엔드 갱신이 실패하면 저장소에 남은 파일을 지운다. 저장소
설정이 비어 있으면 첫 업로드가 아니라 부팅 시점에 실패하도록 했다.

프로필 관련 코드는 전부 src/profile.ts로 옮겼다. index.ts는 라우터와
초기화 함수만 가져다 쓴다.
