+++ src/index.ts
@@ -2,19 +2,13 @@ import cookieParser from "cookie-parser";
 import express, { NextFunction, Request, Response } from "express";
 import rateLimit from "express-rate-limit";
 import i18n from "./i18n";
-import multer from "multer";
-import path from "path";
 import fetch from "node-fetch";
 import { exec } from "child_process";
-import * as tf from "@tensorflow/tfjs";
-import "@tensorflow/tfjs-backend-wasm";
-import * as nsfw from "nsfwjs";
-import fs from "fs";
-import sharp from "sharp";
 import { logger } from "./logger";
 import { errorHandler, notFoundHandler, sendError } from "./middleware";
+import { initProfile, profileRouter } from "./profile";
 import { URL } from "url";
-import { createHash, randomUUID } from "crypto";
+import { createHash } from "crypto";
 
 let branch;
 exec("git branch --show-current", (err, stdout, stderr) => {
@@ -29,15 +23,13 @@ const config = require(__dirname + "/../config/config.json");
 
 const version = require(__dirname + "/../package.json").version;
 
-let model;
-
 // node-fetch has no default timeout: a hung backend would pin the request handler.
 const API_TIMEOUT_MS = 5000;
 
 const app = express();
 app.locals.pretty = true;
 
-// 버전 노출을 막습니다.
+// Do not advertise the framework version.
 app.disable("x-powered-by");
 
 // Rate limiting keys on the client address, which arrives via X-Forwarded-For.
@@ -138,18 +130,6 @@ const gateLimiter = rateLimit({
   windowMs: 60 * 1000,
   limit: 30,
   skip: (req) => !!req.headers.cookie && readCachedStatus(authCacheKey(req.headers.cookie)) !== null,
-  standardHeaders: true,
-  legacyHeaders: false,
-});
-
-/**
- * An upload costs a 3MB write, a sharp re-encode and an NSFW inference, and the
- * file it leaves behind is served from our own origin. Cheap for the caller,
- * expensive for us -- so it gets a budget of its own, well below the gate's.
- */
-const uploadLimiter = rateLimit({
-  windowMs: 60 * 1000,
-  limit: 10,
   standardHeaders: true,
   legacyHeaders: false,
 });
@@ -318,309 +298,12 @@ app.get("/privacy", (req, res) => {
   res.render("privacy");
 });
 
-// 업로드된 프로필 이미지가 놓이는 유일한 위치입니다. 아래 세 곳이 각자
-// 다른 방식으로 같은 경로를 만들고 있어 하나로 모았습니다.
-const PROFILES_DIR = path.resolve(__dirname, "..", "public", "images", "profiles");
+app.use(profileRouter);
 
-/**
- * 프로필 디렉터리 "안의 파일"일 때만 정규화한 절대 경로를 돌려줍니다. 벗어나거나
- * 디렉터리 자기 자신이면 null입니다. 삭제처럼 되돌릴 수 없는 동작 앞에 꼭
- * 거쳐야 합니다.
- */
-const insideProfiles = (candidate: string): string | null => {
-  const resolved = path.resolve(PROFILES_DIR, candidate);
-  return resolved.startsWith(PROFILES_DIR + path.sep) ? resolved : null;
-};
-
-const upload = multer({
-  storage: multer.diskStorage({
-    destination: function (req, file, cb) {
-      const dirPath = PROFILES_DIR;
-      fs.mkdir(dirPath, { recursive: true }, (err) => {
-        if (err) {
-          logger.error("Profile update API error", err, {
-            message: "Failed to create profile image directory.",
-            userid: req.params.userid,
-          });
-          return cb(err, dirPath);
-        }
-        cb(null, dirPath);
-      });
-    },
-    // Date.now() alone collides: two uploads in the same millisecond overwrite
-    // each other's image.
-    filename: function (req, file, cb) {
-      cb(null, `${Date.now()}-${randomUUID()}.webp`);
-    },
-  }),
-  limits: {
-    fileSize: 3 * 1024 * 1024, // 3MB
-  },
-  // Rejecting here keeps a non-image from ever reaching the disk; the handler's
-  // own check runs after multer has already written the file.
-  fileFilter: function (req, file, cb) {
-    cb(null, file.mimetype.indexOf("image") === 0);
-  },
-}).single("img");
-
-// multer writes the upload to disk before the handler runs, so every path that
-// rejects the request afterwards has to remove it or the bytes stay forever.
-const discardUpload = (file: Express.Multer.File | undefined) => {
-  if (!file?.path) return;
-  // 파일명은 서버가 만들지만, 지우는 동작이므로 경로를 다시 확인합니다.
-  const target = insideProfiles(file.path);
-  if (!target) {
-    logger.error("Refused to remove a file outside the profiles directory", null, { path: file.path });
-    return;
-  }
-  fs.promises.unlink(target).catch((err) => {
-    if (err.code !== "ENOENT") logger.warn("Failed to remove a rejected upload", { path: target, error: err });
-  });
-};
-
-const imageToTensor = async (fileBuffer) => {
-  const { data, info } = await sharp(fileBuffer).raw().toBuffer({ resolveWithObject: true });
-
-  return tf.tensor3d(new Uint8Array(data), [info.height, info.width, info.channels], "int32");
-};
-
-/**
- * Resolve the caller's identity from the backend session.
- *
- * This route supplies the backend's project secret on the caller's behalf, so
- * the `:userid` path segment must never be trusted. The browser already sends
- * its session cookie here, so forward it and let the backend session decide.
- */
-const resolveSessionUserid = async (req): Promise<string | null> => {
-  if (!req.headers.cookie) return null;
-  try {
-    const response = await fetch(`${config.project.api}/user`, {
-      method: "GET",
-      headers: {
-        "Content-Type": "application/json",
-        Cookie: req.headers.cookie,
-      },
-      signal: AbortSignal.timeout(API_TIMEOUT_MS),
-    });
-    if (!response.ok) return null;
-    const data: any = await response.json();
-    if (data.result !== "success" || !data.user || !data.user.userid) return null;
-    return String(data.user.userid);
-  } catch (err) {
-    logger.warn("Failed to resolve session userid", { error: err });
-    return null;
-  }
-};
-
-app.post("/profile/:userid/:type", uploadLimiter, async (req, res) => {
-  // Validate userid: must be numeric
-  if (!/^[0-9]+$/.test(req.params.userid)) {
-    res.status(400).json({
-      result: "failed",
-      message: "Invalid userid format",
-      error: "Bad userid",
-    });
-    return;
-  }
-
-  // Authenticate before multer, sharp and the NSFW model spend any work.
-  const sessionUserid = await resolveSessionUserid(req);
-  if (!sessionUserid) {
-    res.status(401).json({
-      result: "failed",
-      message: "Login required",
-      error: "Unauthorized",
-    });
-    return;
-  }
-  if (sessionUserid !== req.params.userid) {
-    logger.warn("Rejected profile upload for a different user", {
-      sessionUserid,
-      requested: req.params.userid,
-    });
-    res.status(403).json({
-      result: "failed",
-      message: "You can only update your own profile",
-      error: "Forbidden",
-    });
-    return;
-  }
-
-  let type = "";
-  let width = 256;
-  let height = 256;
-  if (req.params.type == "picture") {
-    type = "picture";
-  } else if (req.params.type == "background") {
-    width = 2560;
-    height = null;
-    type = "background";
-  } else {
-    res.status(400).json({
-      result: "failed",
-      message: "Error occured while uploading",
-      error: "Invalid type",
-    });
-    return;
-  }
-
-  // Fetch existing profile/background URL before upload
-  let oldFileUrl: string | null = null;
-  try {
-    const profileResponse = await fetch(`${config.project.api}/profile/${req.params.userid}`, {
-      method: "GET",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      signal: AbortSignal.timeout(API_TIMEOUT_MS),
-    });
-    const profileData: any = await profileResponse.json();
-    if (profileData.result === "success" && profileData.user && profileData.user[type] && !config.project.ignoredImageURL.some((domain) => profileData.user[type].includes(domain))) {
-      oldFileUrl = profileData.user[type];
-    }
-  } catch (err) {
-    logger.warn("Failed to fetch existing profile data", { userid: req.params.userid, type, error: err });
-  }
-
-  upload(req, res, async (err) => {
-    if (err) {
-      if (err instanceof multer.MulterError) err = err.message;
-      else err = err.code;
-      discardUpload(req.file);
-      logger.error("File upload error", err, { userid: req.params.userid, type: req.params.type });
-      res.status(400).json({
-        result: "failed",
-        message: "Error occured while uploading",
-        error: err,
-      });
-      return;
-    }
-    const file = req.file;
-    if (!file || file.mimetype.indexOf("image") == -1) {
-      discardUpload(file);
-      logger.warn("Invalid file upload attempt", { userid: req.params.userid, type: req.params.type, mimetype: file?.mimetype });
-      res.status(400).json({
-        result: "failed",
-        message: "Error occured while uploading",
-        error: "Invalid file type",
-      });
-      return;
-    }
-    const filePath = insideProfiles(file.path);
-    if (!filePath) {
-      discardUpload(file);
-      logger.error("Path traversal attempt detected", null, { userid: req.params.userid, path: file.path });
-      res.status(400).json({
-        result: "failed",
-        message: "Error occured while uploading",
-        error: "Invalid file path",
-      });
-      return;
-    }
-
-    // multer's callback is not an Express handler, so a rejection here is not
-    // turned into a response -- the request would hang and the file would stay.
-    let fileBuffer: Buffer;
-    let explicit = false;
-    try {
-      fileBuffer = await sharp(filePath)
-        .resize({ width, height })
-        .flatten({ background: "#ffffff" })
-        .webp({
-          quality: 70,
-          effort: 6,
-        })
-        .toBuffer();
-      fs.writeFileSync(filePath, fileBuffer);
-      const image = await imageToTensor(fileBuffer);
-      const predictions = await model.classify(image);
-      image.dispose();
-      if (predictions[0].className != "Drawing" && predictions[0].className != "Neutral") explicit = true;
-    } catch (e) {
-      discardUpload(file);
-      logger.error("Failed to process the uploaded image", e, { userid: req.params.userid, type: req.params.type });
-      res.status(400).json({
-        result: "failed",
-        message: "Error occured while uploading",
-        error: "Invalid image",
-      });
-      return;
-    }
-
-    fetch(`${config.project.api}/profile/${type}`, {
-      method: "PUT",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      signal: AbortSignal.timeout(API_TIMEOUT_MS),
-      body: JSON.stringify({
-        explicit,
-        // Verified against the backend session, never the raw path segment.
-        userid: sessionUserid,
-        value: `${config.project.url}/images/profiles/${file.filename}`,
-        secret: config.project.secretKey,
-      }),
-    })
-      .then((res) => res.json())
-      .then((json: any) => {
-        if (json.result == "failed") {
-          discardUpload(file);
-          logger.error("Profile update API error", null, { message: json.message, userid: req.params.userid });
-          res.status(400).json({
-            result: "failed",
-            message: "Error occured while uploading",
-            error: json.message,
-          });
-          return;
-        }
-
-        // Delete old file after successful upload
-        if (oldFileUrl) {
-          // Asynchronously delete the old file without blocking the response.
-          (async () => {
-            try {
-              const oldFilename = path.basename(new URL(oldFileUrl).pathname);
-              const oldFilePath = insideProfiles(oldFilename);
-
-              if (oldFilePath) {
-                await fs.promises.unlink(oldFilePath);
-                logger.info(`Deleted old ${type} file`, { userid: req.params.userid, oldFilename });
-              } else {
-                logger.warn("Skipping deletion of file outside the root directory.", { userid: req.params.userid, oldFileUrl });
-              }
-            } catch (err) {
-              // It's okay if the file doesn't exist. Log other errors.
-              if (err.code !== "ENOENT") {
-                logger.warn(`Failed to delete old ${type} file`, { userid: req.params.userid, oldFileUrl, error: err });
-              }
-            }
-          })();
-        }
-
-        res.status(200).json({ result: "success", url: `${config.project.url}/images/profiles/${file.filename}`, explicit });
-      })
-      .catch((err) => {
-        discardUpload(file);
-        logger.error("Profile update fetch error", err, { userid: req.params.userid });
-        res.status(500).json({
-          result: "failed",
-          message: "Error occured while uploading",
-          error: "Internal server error",
-        });
-      });
-  });
-});
-
-const loadModel = async () => {
-  model = await nsfw.load(`${config.project.nsfw}/mobilenet_v2/`);
-};
-
-// Handle unhandled promise rejections
 process.on("unhandledRejection", (reason: any, promise: Promise<any>) => {
   logger.fatal("Unhandled Promise Rejection", reason, { promise: promise.toString() });
 });
 
-// Handle uncaught exceptions
 process.on("uncaughtException", (error: Error) => {
   logger.fatal("Uncaught Exception", error);
   process.exit(1);
@@ -628,10 +311,7 @@ process.on("uncaughtException", (error: Error) => {
 
 (async () => {
   try {
-    await tf.setBackend("wasm");
-    await tf.ready();
-
-    await loadModel();
+    await initProfile();
 
     app.listen(config.project.port, () => {
       logger.info(`URLATE-v3l-frontend is running on version ${config.project.mode == "test" ? Date.now() : version}.`);
@@ -643,8 +323,8 @@ process.on("uncaughtException", (error: Error) => {
   }
 })();
 
-// 라우트 뒤에 와야 합니다. 앞에 두면 모든 요청이 404로 끝납니다.
+// Must come after the routes: placed earlier, every request would end in a 404.
 app.use(notFoundHandler);
 
-// Add error handler middleware (must be last)
+// Must be last.
 app.use(errorHandler);
